move tailboom flex model to empennage model
